fixed tons and tons of warnings cleaned up macros a little
git-svn-id: https://bugengine.googlecode.com/svn/trunk@53 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/graphics/api/graphics/renderer/indexbuffer.hh
@@ -39,10 +39,10 @@ private:
     const RenderBackend*    m_owner;
     refptr<GpuBuffer>       m_buffer;
 public:
-    IndexBuffer(const RenderBackend* owner, size_t vertexCount, IndexUsage usage);
+    IndexBuffer(const RenderBackend* owner, u32 vertexCount, IndexUsage usage);
     ~IndexBuffer();
 
-    T*  map(size_t count = 0, size_t offset = 0);
+    T*  map(u32 count = 0, u32 offset = 0);
     void unmap();
 
     const GpuBuffer* buffer() const;
